+++ CHANGES.md
@@ -1,3 +1,15 @@
+
+
+
+- Improve `ocamlbuild` cross-compilation support. Adds the
+  `Conf.toolchain` configuration key. If specified on the
+  command line or via the `TOPKG_CONF_TOOLCHAIN` environment
+  variable. It's value is used with the `-toolchain` option
+  introduced in `ocamlbuild` 0.11.0 in default build command
+  `Pkg.build_cmd`. If unspecified the default build command
+  is left unchanged. Thanks to Peter Zotov for the patch.
+
+
 v0.8.1 2016-11-02 Zagreb
 ------------------------